Bookmark Sign In loading spinner (#279)
## Description

When a user clicks the Sign In button in the Bookmark modal, it takes a
few seconds to redirect the user to auth0, this is because it goes
through the middleware and needs to load some js. This PR adds a loading
spinner to the button to let the user know something is loading.

<img width="1234" height="731" alt="image"
src="https://github.com/user-attachments/assets/a2577ffa-866d-4ece-add4-c264f725533b"
/>